requirements.txt
Flask==3.1.1
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
itsdangerous==2.2.0
blinker==1.9.0
click==8.2.1

numpy==2.3.2
pillow==11.3.0

pycryptodome==3.23.0
